express router leveraged to optimize backend
