+++ server/utils/authMiddleware.js
